pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "system-monitor-api"
version = "1.0.0"
description = "A comprehensive system monitoring API built with FastAPI and psutil"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["system", "monitoring", "api", "fastapi", "psutil", "metrics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Monitoring",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/system-monitor"
Documentation = "https://github.com/yourusername/system-monitor#readme"
Repository = "https://github.com/yourusername/system-monitor"
"Bug Reports" = "https://github.com/yourusername/system-monitor/issues"

[project.scripts]
system-monitor = "system_monitor.main:main"
system-monitor-register = "system_monitor.register:register"

